Renamed filenames from kebab-case to camelCase (#2192)
* refactor(boilerplate): reworked kebab-case to camelCase

* refactor(test): move mocks into setup

* refactor(mock-file): camelCase

* docs(boilerplate): updated readme
