+++ boilerplate/README.md
@@ -22,18 +22,18 @@ The Ignite boilerplate project's structure will look similar to this:
 ignite-project
 ├── app
 │   ├── components
+│   ├── config
 │   ├── i18n
-│   ├── utils
 │   ├── models
 │   ├── navigators
 │   ├── screens
 │   ├── services
 │   ├── theme
+│   ├── utils
 │   ├── app.tsx
 ├── test
 │   ├── __snapshots__
-│   ├── mock-i18n.ts
-│   ├── mock-reactotron.ts
+│   ├── mockFile.ts
 │   ├── setup.ts
 ├── README.md
 ├── android
@@ -68,23 +68,24 @@ ignite-project
 
 Included in an Ignite boilerplate project is the `app` directory. This is a directory you would normally have to create when using vanilla React Native.
 
-The inside of the src directory looks similar to the following:
+The inside of the `app` directory looks similar to the following:
 
 ```
 app
-│── components
-│── i18n
+├── components
+├── config
+├── i18n
 ├── models
 ├── navigators
 ├── screens
 ├── services
 ├── theme
 ├── utils
-└── app.tsx
+├── app.tsx
 ```
 
 **components**
-MAVERICKTODO: update this documentation
+This is where your reusable components live which help you build your screens.
 
 **i18n**
 This is where your translations will live if you are using `react-native-i18n`.
@@ -105,7 +106,7 @@ Any services that interface with the outside world will live here (think REST AP
 Here lives the theme for your application, including spacing, colors, and typography.
 
 **utils**
-This is a great place to put miscellaneous helpers and utilities. Things like date helpers, formatters, etc. are often found here. However, it should only be used for things that are truely shared across your application. If a helper or utility is only used by a specific component or model, consider co-locating your helper with that component or model.
+This is a great place to put miscellaneous helpers and utilities. Things like date helpers, formatters, etc. are often found here. However, it should only be used for things that are truly shared across your application. If a helper or utility is only used by a specific component or model, consider co-locating your helper with that component or model.
 
 **app.tsx** This is the entry point to your app. This is where you will find the main App component which renders the rest of the application.
 
@@ -126,3 +127,4 @@ Read [Detox setup instructions](./detox/README.md).
 - [2018 aka Bowser](https://github.com/infinitered/ignite-bowser)
 - [2017 aka Andross](https://github.com/infinitered/ignite-andross)
 - [2016 aka Ignite 1.0](https://github.com/infinitered/ignite-ir-boilerplate-2016)
+
