+++ boilerplate/app/app.tsx
@@ -10,7 +10,7 @@
  * if you're interested in adding screens and navigators.
  */
 import "./i18n"
-import "./utils/ignore-warnings"
+import "./utils/ignoreWarnings"
 import { useFonts } from "expo-font"
 import React from "react"
 import { initialWindowMetrics, SafeAreaProvider } from "react-native-safe-area-context"
