+++ boilerplate/app/models/helpers/useStores.ts
@@ -1,7 +1,7 @@
 import { createContext, useContext, useEffect, useState } from "react"
 import { setReactotronRootStore } from "../../services/reactotron"
 import { RootStore, RootStoreModel } from "../RootStore"
-import { setupRootStore } from "./setup-root-store"
+import { setupRootStore } from "./setupRootStore"
 
 /**
  * Create the initial (empty) global RootStore instance here.
