+++ boilerplate/app/utils/crashReporting.ts
