+++ boilerplate/app/utils/openLinkInBrowser.ts
