--- boilerplate/test/mock-async-storage.ts
@@ -1,3 +1,0 @@
-import mockAsyncStorage from "@react-native-async-storage/async-storage/jest/async-storage-mock"
-
-jest.mock("@react-native-async-storage/async-storage", () => mockAsyncStorage)
