--- boilerplate/test/mock-reactotron.ts
@@ -1,1 +1,0 @@
-declare const tron // eslint-disable-line @typescript-eslint/no-unused-vars
