+++ boilerplate/test/setup.ts
@@ -1,11 +1,39 @@
 // we always make sure 'react-native' gets included first
-import "react-native"
+import * as ReactNative from "react-native"
+import mockAsyncStorage from "@react-native-async-storage/async-storage/jest/async-storage-mock"
+import mockFile from "./mockFile"
 
 // libraries to mock
-import "./mock-react-native-image"
-import "./mock-async-storage"
-import "./mock-i18n"
-import "./mock-reactotron"
+jest.doMock("react-native", () => {
+  // Extend ReactNative
+  return Object.setPrototypeOf(
+    {
+      Image: {
+        ...ReactNative.Image,
+        resolveAssetSource: jest.fn((_source) => mockFile), // eslint-disable-line @typescript-eslint/no-unused-vars
+        getSize: jest.fn(
+          (
+            uri: string, // eslint-disable-line @typescript-eslint/no-unused-vars
+            success: (width: number, height: number) => void,
+            failure?: (_error: any) => void, // eslint-disable-line @typescript-eslint/no-unused-vars
+          ) => success(100, 100),
+        ),
+      },
+    },
+    ReactNative,
+  )
+})
+
+jest.mock("@react-native-async-storage/async-storage", () => mockAsyncStorage)
+
+jest.mock("i18n-js", () => ({
+  currentLocale: () => "en",
+  t: (key: string, params: Record<string, string>) => {
+    return `${key} ${JSON.stringify(params)}`
+  },
+}))
+
+declare const tron // eslint-disable-line @typescript-eslint/no-unused-vars
 
 jest.useFakeTimers()
 declare global {
